# documentacao/swagger.yaml
openapi: 3.0.3
info:
  title: FinBertoldi API
  description: |
    API do sistema de controle financeiro familiar FinBertoldi.

    A aplicacao principal (porta 8080) serve HTML via server-side rendering.
    As rotas POST sao form submissions que retornam redirects (303).

    O pluggy-service (porta 8081) expoe uma API REST JSON interna.
  version: 1.0.0
  contact:
    name: FinBertoldi
servers:
  - url: http://localhost:8080
    description: App principal (desenvolvimento)
  - url: http://localhost:8081
    description: Pluggy Service (desenvolvimento)
  - url: http://<SERVER_IP>
    description: Producao (ver PRODUCAO.md para IP)
...
paths:
...
  /api/pluggy/connect-token:
    post:
      tags: [Pluggy Service]
      summary: Gerar connect token Pluggy
      servers:
        - url: http://localhost:8081
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                clientUserId:
                  type: string
      responses:
        '200':
          description: Token gerado
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
  /api/pluggy/webhook:
    post:
      tags: [Pluggy Service]
      summary: Webhook da Pluggy
      description: Recebe eventos item/created e item/updated
      servers:
        - url: http://localhost:8081
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                event:
                  type: string
                  enum: [item/created, item/updated]
                id:
                  type: string
                  description: ID do item
      responses:
        '200':
          description: Webhook processado
  /api/pluggy/items:
    post:
      tags: [Pluggy Service]
      summary: Registrar item conectado
      servers:
        - url: http://localhost:8081
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [itemId, familyId]
              properties:
                itemId:
                  type: string
                familyId:
                  type: integer
                connectorName:
                  type: string
      responses:
        '200':
          description: Item registrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /api/pluggy/sync/{item_id}:
    post:
      tags: [Pluggy Service]
      summary: Forcar sync de um item
      servers:
        - url: http://localhost:8081
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sync concluido
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced:
                    type: integer
                    description: Numero de novas transacoes
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sessao HTTP-only (token hex 64 chars)